添加更新录入配置功能，优化下载与解析并集成到系统

+++ coffee/Utils.coffee
@@ -11,6 +11,7 @@ Utils.rmdir = (path, callback)->
 		files = fs.readdirSync path
 		files.forEach (file, index)->
 			curPath = path + "/" + file
+			return if !fs.existsSync curPath
 			if fs.statSync(curPath).isDirectory()
 				Utils.rmdir curPath, ()->
 			else
